% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trolMail.R
\name{trolMail}
\alias{trolMail}
\title{Function that prepares & sends a friendly email for you.}
\usage{
trolMail()
}
\description{
Function that prepares & sends a friendly email for you.
}
\examples{
trolMail()

}
\author{
Emelie Hofland, \email{emelie_hofland@hotmail.com}
}
